% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{baseline_correct}
\alias{baseline_correct}
\title{Baseline correct}
\usage{
baseline_correct(data)
}
\arguments{
\item{data}{dataframe Usually from \code{\link{read_data}}. Must contain columns called
"Intensity" and "Time".}
}
\value{
dataframe with corrected intensities
}
\description{
Baseline correct the raw intensity data.  Uses \code{\link[baseline]{baseline}} with the "irls"
method to correct the baseline.
}

