% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{detect_peaks}
\alias{detect_peaks}
\title{Detect peaks}
\usage{
detect_peaks(data, updateProgress = NULL, ...)
}
\arguments{
\item{data}{dataframe Usually from \code{\link{baseline_correct}}. Must contain columns called
"Intensity" and "Time".}

\item{...}{Other parameters passed to findpeaks}
}
\value{
dataframe with information about each peak
}
\description{
The main function to detect the peaks in the data.  Best results with baseline corrected data.
Uses the \code{\link[pracma]{findpeaks}} function.
}
