% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{filter_by_tracks}
\alias{filter_by_tracks}
\title{Filter by tracks}
\usage{
filter_by_tracks(data, track_size_cutoff)
}
\arguments{
\item{track_size_cutoff}{}
}
\description{
Filter data by the number of tracks or time points.  Used to get rid of incomplete tracks.
}

