% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stats.R
\name{get_stats}
\alias{get_stats}
\title{Get well stats}
\usage{
get_stats(data, peaks)
}
\arguments{
\item{data}{intensity data}

\item{peaks}{peak data from \code{\link{detect_peaks}}}
}
\value{
a data.frame with the data summarized by well.  Contains columns for the number of
  tracks, number of events and stats on the peak width and number of peaks.
}
\description{
Get stats about peaks present in each well
}

