% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_tracks}
\alias{plot_tracks}
\title{Plot individual tracks}
\usage{
plot_tracks(data, well = "All", y = "Intensity", peak_data = NULL,
  filter_peaks = FALSE, show_peaks = TRUE)
}
\arguments{
\item{data}{The raw data, usually baseline corrected with \code{\link{baseline_correct}}}

\item{well}{The well to plot}

\item{y}{The name of the y-axis}

\item{peak_data}{For showing peaks, generated by \code{\link{detect_peaks}}}

\item{filter_peaks}{logical Show only spots with detected peaks}

\item{show_peaks}{logical Show detected peaks in plot}
}
\value{
a ggplot object
}
\description{
Plot the individual tracks of a well with the option to show peaks and to show only tracks with
detected peaks
}

