% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_width_of_peaks}
\alias{plot_width_of_peaks}
\title{Plot width of peaks}
\usage{
plot_width_of_peaks(peak_stats, peaks)
}
\arguments{
\item{peak_stats}{A peaks stats data frame generated by \code{\link{get_stats}}}

\item{peaks}{The dataframe with detected peaks generated by \code{\link{detect_peaks}}}
}
\description{
Plot the average width of peaks detected per well
}

