% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_data}
\alias{read_data}
\title{Read the raw data into memory}
\usage{
read_data(path, section = "Calibration", convert_time = 1000 * 60,
  channel = "Ch 3")
}
\arguments{
\item{path}{File path of the csv file}

\item{section}{Name of the section to import. Default is 'Calibration'.}

\item{convert_time}{Conversion factor for the time column.  Default is \eqn{1000 * 60} (converts
milliseconds to minutes). NULL means no conversion will be performed.}

\item{channel}{The channel to use for intensity measurements. Default is 'Ch 3'}
}
\value{
dataframe
}
\description{
This function takes the raw csv files exported by the imaging software and loads it into memory.
It is important that this is the csv file directly exported by the program and not modified in
any way.  It detectes the correct section of the csv file to import by looking for the
'========>Calibration<========' flag.  It will import from this flag to either the end of the
file or the next flag, which ever comes first.  It also cleans up the raw data, removing
extraneous columns and extracting the well name and multiple fields of view if present.
}

