% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{scan_file}
\alias{scan_file}
\title{Scan file}
\usage{
scan_file(path)
}
\arguments{
\item{path}{file path of the csv file}
}
\value{
list Returns a named list with two entries, channels and sections
}
\description{
Scan raw file for channels and sections
}
